+++ src/core/diffusion/runge_kutta_grid.h
@@ -34,7 +34,7 @@ namespace bdm {
 class RungeKuttaGrid : public DiffusionGrid {
  public:
   RungeKuttaGrid() = default;
-  RungeKuttaGrid(int substance_id, std::string substance_name, double dc,
+  RungeKuttaGrid(size_t substance_id, std::string substance_name, double dc,
                  int resolution = 10)
       : DiffusionGrid(substance_id, std::move(substance_name), dc, 0.0,
                       resolution) {}
